refactor: major refactor in small modules
