+++ polus/_plugins/_plugin_model.py
@@ -4,7 +4,6 @@
 
 from __future__ import annotations
 
-from enum import Enum
 from typing import Any, List, Optional, Union
 
 from pydantic import BaseModel, Field, constr, validator
